feat(IM业务): 心跳机制基础
（1）添加心跳机制 （2）心跳时间

+++ src/main/java/com/howcode/aqchat/config/AQChatConfig.java
@@ -17,5 +17,5 @@ public class AQChatConfig {
     private int webSocketPort;
     private int bossThreadSize;
     private int workThreadSize;
-    private int heartBeatTime;
+    private Long heartBeatTime;
 }
